+++ Frontend/src/App.tsx
@@ -1,28 +1,20 @@
 import { Routes, Route } from 'react-router-dom';
 import NotFound from './Components/Pages/NotFound';
-import TimelinePage from './Components/Pages/TimelinePage';
+import TimelinePage from './Components/Pages/PublicTimelinePage';
 import SnackbarContextProvider from './Components/SnackBar/SnackbarContextProvider';
 import Snackbar from './Components/SnackBar/Snackbar';
-import MyTimeline from './Components/Pages/MyTimeline';
+import MyTimelinePage from './Components/Pages/MyTimelinePage';
+import UserTimelinePage from './Components/Pages/UserTimelinePage';
 
 function App() {
-	// const { isLoading } = useAuth0();
-
-	// if (isLoading) {
-	// 	return (
-	// 		<div className="page-layout">
-	// 			<PageLoader />
-	// 		</div>
-	// 	);
-	// }
 
 	return (
 		<SnackbarContextProvider>
 			<Routes>
-				<Route path="/" element={<TimelinePage />}></Route>
-				<Route path="user/:username" element={<TimelinePage />} />
+				<Route path="/" element={<TimelinePage />}/>
+				<Route path="user/:username" element={<UserTimelinePage />} />
 				<Route path="*" element={<NotFound />} />
-				<Route path=":username/my-timeline" element={<MyTimeline/>}/>
+				<Route path=":username/my-timeline" element={<MyTimelinePage/>}/>
 			</Routes>
 			<Snackbar />
 		</SnackbarContextProvider>
