+++ Frontend/src/Components/Pages/PublicTimelinePage.tsx
@@ -1,0 +1,110 @@
+import React, { FC, useEffect, useState } from 'react';
+import { Box } from '@mui/material';
+import Header from '../Global/Header/Header';
+import { Message } from '../../Types/Timeline';
+import TimeLineContainer from '../PublicTimeline/TimeLineContainer';
+import { isLoggedIn } from '../Authentication/cookieHandler';
+import { IsLoading } from '../../Types/Global';
+import axios, { AxiosError } from 'axios';
+import Tweet from '../Tweet/Tweet';
+
+
+
+const PublicTimelinePage: FC = () => {
+	const style = {
+		container: {
+			display: 'flex',
+			flexDirection: 'column',
+			alignItems: 'center',
+			justifyContent: 'center',
+		},
+	};
+
+	const [timeline, setTimeline] = useState([] as Message[]);
+
+	const [hasMore, setHasMore] = useState(true);
+
+	const [isLoading, setIsLoading] = useState({
+		isLoading: true,
+		error: null,
+	} as IsLoading);
+
+	const getTimeline = async (startIndex?: number, endIndex?: number) => {
+		const baseUrl = `${process.env.REACT_APP_API_SERVER_URL}/msgs`;
+		const queryParams = startIndex !== undefined && endIndex !== undefined ? `?startIndex=${startIndex}&endIndex=${endIndex}` : '';
+
+		const options =
+			{
+				headers: {
+					'access-control-allow-origin': `${process.env.REACT_APP_API_SERVER_URL}`,
+				},
+			};
+
+		const fullUrl = baseUrl + queryParams;
+
+		axios
+			.get(fullUrl, options)
+			.then((res) => {
+				setTimeline(res.data);
+				setIsLoading({ isLoading: false, error: null });
+			})
+			.catch((e) => {
+				if (e instanceof AxiosError) {
+					console.log(e);
+					setIsLoading({ isLoading: false, error: e.message });
+				} else {
+					console.log(e);
+					setIsLoading({ isLoading: false, error: 'Something went wrong!' });
+				}
+			});
+	};
+
+	const fetchMoreData = () => {
+		const startIndex = timeline.length;
+		const endIndex = startIndex + 40;
+
+		const baseUrl = `${process.env.REACT_APP_API_SERVER_URL}/msgs`;
+
+		const queryParams = startIndex !== undefined && endIndex !== undefined ? `?startIndex=${startIndex}&endIndex=${endIndex}` : '';
+
+		const fullUrl = baseUrl + queryParams;
+
+		const options =
+			{
+				headers: {
+					'access-control-allow-origin': `${process.env.REACT_APP_API_SERVER_URL}`,
+				},
+			}
+
+		axios
+			.get(fullUrl, options)
+			.then((res) => {
+				console.log(res.data);
+				if (res.data.length === 0) {
+					setHasMore(false);
+				}
+				setTimeline([...timeline, ...res.data]);
+			})
+			.catch((e) => {
+				if (e instanceof AxiosError) {
+					console.log(e);
+				} else {
+					console.log(e);
+				}
+			});
+	};
+
+	useEffect(() => {
+		getTimeline(0, 40);
+	}, []);
+	
+	return (
+			<Box sx={style.container}>
+				<Header title="Public Timeline"/>
+				<Tweet updateTweetsCallback={getTimeline} />
+				<TimeLineContainer messages={timeline} getNextMessages={fetchMoreData} hasMore={hasMore} isLoading={isLoading}/>
+			</Box>
+	);
+};
+
+export default PublicTimelinePage;
